move payments in new file
